requirements.txt
click<8
stevedore
jenkins-job-builder>=2.0.0.0b1
python-jenkins>=0.4.15
